requirements.txt
django
dj-database-url
Unipath
gunicorn
psycopg2
django-mptt
